+++ activestorage/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Support multiple analyzers during ActiveStorage::Blob#analyze`
+
+    *Sean Doyle*
 
 Please check [7-1-stable](https://github.com/rails/rails/blob/7-1-stable/activestorage/CHANGELOG.md) for previous changes.